% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsts-state-matrices.R
\name{bsts_full_states}
\alias{bsts_full_states}
\alias{bsts_final_state}
\alias{bsts_posterior}
\title{BSTS Extractor Functions}
\usage{
bsts_full_states(object)

bsts_final_state(object)

bsts_posterior(object)
}
\arguments{
\item{object}{object of class \dQuote{bsts.estimate}.}
}
\description{
BSTS Extractor Functions
}
\details{
The full and final states functions extract the smoothed states 
distribution from the BSTS estimated object, and the posterior function the
draws x parameters matrix.
}
