\name{bsts_full_states}
\alias{bsts_full_states}
\title{
Extracts the BSTS full states distribution.
}
\description{
Extracts the full states distribution from the BSTS object.
}
\usage{
bsts_full_states(object)
}
\arguments{
\item{object}{
An object of class \dQuote{bsts.estimate}
}
}
\details{
These are the distributions of the smoothed states. 
}
\value{
A draws x states x time array.
}
\author{
Alexios Galanos
}

