% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsts-model.R
\name{bsts_modelspec}
\alias{bsts_modelspec}
\title{BSTS Model Secification}
\usage{
bsts_modelspec(
  y,
  xreg = NULL,
  frequency = NULL,
  differences = 0,
  level = TRUE,
  slope = TRUE,
  damped = FALSE,
  seasonal = FALSE,
  seasonal_frequency = 4,
  ar = FALSE,
  ar_max = 1,
  cycle = FALSE,
  cycle_frequency = NULL,
  cycle_names = NULL,
  seasonal_type = "regular",
  seasonal_harmonics = NULL,
  transformation = "box-cox",
  lambda = NULL,
  lower = 0,
  upper = 1,
  distribution = "gaussian",
  ...
)
}
\arguments{
\item{y}{an xts vector.}

\item{xreg}{an xts matrix of external regressors.}

\item{frequency}{frequency of y (if using a seasonal model).}

\item{differences}{number of differences to apply to the outcome variable y 
(max of 2).}

\item{level}{whether to include a level component (Local Level Model).}

\item{slope}{whether to include a slope component (Local Linear Model).}

\item{damped}{whether to include a damped trend (damped Local Linear Model).}

\item{seasonal}{whether to include a seasonal component.}

\item{seasonal_frequency}{vector of seasonal frequencies.}

\item{ar}{whether to include a sparse AR component.}

\item{ar_max}{number of lags for the AR component.}

\item{cycle}{whether to include a cyclical component.}

\item{cycle_frequency}{number of periods in a cycle. This can be a vector in 
which case multiple cycles are included.}

\item{cycle_names}{optional vector of cycle names.}

\item{seasonal_type}{type of seasonality (regular or trigonometric).}

\item{seasonal_harmonics}{number of harmonics to include in the seasonal 
component when seasonal_type is trigonometric.}

\item{transformation}{a valid transformation for y from the \dQuote{tstransform} 
function in the \dQuote{tsaux} package (currently box-cox or logit are available).}

\item{lambda}{the Box Cox lambda. If NA will estimate this using the method 
of Guerrero.}

\item{lower}{lower bound for the transformation.}

\item{upper}{upper bound for the transformation.}

\item{distribution}{valid choices are currently only \dQuote{gaussian}.}

\item{...}{not currently used.}
}
\value{
An object of class \dQuote{bsts.spec}.
}
\description{
Specifies a BSTS model prior to estimation.
}
\note{
This is a wrapper to part of the functionality of the bsts package. Once 
an object is estimated, all other methods are implemented locally (including 
prediction).
}
