\name{bsts_posterior}
\alias{bsts_posterior}
\title{
BSTS Posterior Parameters
}
\description{
Extracts the posterior distribution of the BSTS model parameters.
}
\usage{
bsts_posterior(object)
}
\arguments{
\item{object}{
An object of class \dQuote{bsts.estimate}
}
}
\value{
A named draws x parameters matrix.
}
\author{
Alexios Galanos
}
