% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-model.R, R/bsts-model.R
\name{plot.arima.estimate}
\alias{plot.arima.estimate}
\alias{plot}
\alias{plot.bsts.estimate}
\title{Object Plots}
\usage{
\method{plot}{arima.estimate}(x, y = NULL, ...)

\method{plot}{bsts.estimate}(x, y = NULL, ...)
}
\arguments{
\item{x}{an object of class \dQuote{bsts.estimate} or \dQuote{arima.estimate}.}

\item{y}{not used.}

\item{...}{no currently used.}
}
\description{
Plots for objects generated from the bsts or arima functions.
}
