% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsforeign-package.R
\docType{package}
\name{tsforeign-package}
\alias{tsforeign}
\alias{tsforeign-package}
\title{tsforeign: Interface To Other Packages}
\description{
Wrappers for models in other packages to conform to the unified time series framework.
}
\author{
\strong{Maintainer}: Alexios Galanos \email{alexios@4dscape.com}

Other contributors:
\itemize{
  \item James Nesbit \email{jmcgnesbit@gmail.com} [contributor]
  \item Keith O'Hara \email{kth.ohr@gmail.com} [contributor]
}

}
\keyword{internal}
