% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-model.R, R/bsts-model.R
\name{tsmetrics.arima.estimate}
\alias{tsmetrics.arima.estimate}
\alias{tsmetrics}
\alias{tsmetrics.arima.predict}
\alias{tsmetrics.bsts.estimate}
\alias{tsmetrics.bsts.predict}
\title{Performance Metrics}
\usage{
\method{tsmetrics}{arima.estimate}(object, ...)

\method{tsmetrics}{arima.predict}(object, actual, alpha = 0.1, ...)

\method{tsmetrics}{bsts.estimate}(object, ...)

\method{tsmetrics}{bsts.predict}(object, actual, alpha = 0.1, ...)
}
\arguments{
\item{object}{An object of class \dQuote{arima.estimate}, 
\dQuote{arima.predict}, \dQuote{bsts.estimate} or \dQuote{bsts.predict}.}

\item{...}{Optional arguments passed to the MASE function which includes 
\dQuote{frequency}, else will be read from the object (act as an override for 
instance when using fourier seasonality in xreg for arima).}

\item{actual}{the actual data matched to the dates of the forecasts.}

\item{alpha}{the coverage level for distributional forecast metrics.}
}
\description{
Performance metrics from an estimated or predicted model.
}
