% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-model.R, R/bsts-model.R
\name{fitted.arima.estimate}
\alias{fitted.arima.estimate}
\alias{fitted}
\alias{fitted.bsts.estimate}
\title{Model Fitted Values}
\usage{
\method{fitted}{arima.estimate}(object, raw = FALSE, ...)

\method{fitted}{bsts.estimate}(
  object,
  distribution = FALSE,
  invdiff = TRUE,
  raw = FALSE,
  type = c("filtered", "smoothed"),
  ...
)
}
\arguments{
\item{object}{an object of class \dQuote{arima.estimate} or 
\dQuote{bsts.estimate}.}

\item{raw}{if a transformation was used, whether to return the transformed 
fitted values (raw).}

\item{...}{not currently used.}

\item{distribution}{whether to return the full distribution for the bsts 
model.}

\item{invdiff}{if a model in differences was estimated, whether to inverse 
the differencing when calculating the fitted values.}

\item{type}{for the bsts model, the filtered are the one step ahead forecast 
values, whilst the smoothed represent those values using the whole information 
set.}
}
\value{
For the distribution choice, an object of class 
\dQuote{tsmodel.distribution}, else an xts vector.
}
\description{
Extract the fitted values from an estimated model.
}
