\name{plot}
\alias{plot}
\alias{plot.bsts.estimate}
\alias{plot.arima.estimate}
\title{
Model Plots
}
\description{
Plots for objects generated from the bsts or arima functions.
}
\usage{
\method{plot}{bsts.estimate}(x, y = NULL, ...)
\method{plot}{arima.estimate}(x, y = NULL, ...)
}
\arguments{
\item{x}{
An object of class \dQuote{bsts.estimate} or \dQuote{arima.estimate}.
}
\item{y}{
Not used.
}
\item{\dots}{
Additional arguments passed to the underlying plot function.
}
}
\author{
Alexios Galanos
}
