% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-model.R, R/bsts-model.R
\name{residuals.arima.estimate}
\alias{residuals.arima.estimate}
\alias{residuals}
\alias{residuals.bsts.estimate}
\title{Model Residuals}
\usage{
\method{residuals}{arima.estimate}(object, raw = FALSE, ...)

\method{residuals}{bsts.estimate}(
  object,
  distribution = FALSE,
  invdiff = TRUE,
  standardize = FALSE,
  raw = FALSE,
  type = c("filtered", "smoothed"),
  ...
)
}
\arguments{
\item{object}{an object of class \dQuote{tsissm.estimate}.}

\item{raw}{raw residuals are the model based values in transformed space 
(when either Box Cox or Logistic have been used as transformations).}

\item{...}{not currently used.}

\item{distribution}{whether to return the full distribution for the bsts 
model.}

\item{invdiff}{if a model in differences was estimated, whether to inverse 
the differencing when calculating the residual values.}

\item{standardize}{whether to scale the errors by the model standard 
deviation.}

\item{type}{for the bsts model, the filtered are the one step ahead forecast 
errors, whilst the smoothed represent those values using the whole 
information set.}
}
\value{
For the distribution choice, an object of class 
\dQuote{tsmodel.distribution}, else an xts vector.
}
\description{
Extract the residual values from an estimated model.
}
