% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-model.R, R/bsts-model.R
\name{summary.arima.estimate}
\alias{summary.arima.estimate}
\alias{summary}
\alias{summary.bsts.estimate}
\title{Model Estimation Summary}
\usage{
\method{summary}{arima.estimate}(object, ...)

\method{summary}{bsts.estimate}(object, quantiles = c(0.025, 0.25, 0.5, 0.75, 0.975), ...)
}
\arguments{
\item{object}{an object of class \dQuote{arima.spec} or \dQuote{bsts.spec}.}

\item{...}{not currently used.}

\item{quantiles}{vector of quantiles to evaluate for each variable.}
}
\value{
A printout of the parameter summary, model type and some model metrics. 
For the bsts model, the posterior distribution of parameters is fist converted 
into an \link{mcmc} object (coda package).
}
\description{
Summary method.
}
