% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima-model.R, R/bsts-model.R
\name{tsbacktest.arima.spec}
\alias{tsbacktest.arima.spec}
\alias{tsbacktest}
\alias{tsbacktest.bsts.spec}
\title{Walk Forward Model Backtest}
\usage{
\method{tsbacktest}{arima.spec}(
  object,
  start = floor(NROW(object$target$y_orig)/2),
  end = NROW(object$target$y_orig),
  h = 1,
  alpha = NULL,
  trace = FALSE,
  ...
)

\method{tsbacktest}{bsts.spec}(
  object,
  start = floor(NROW(object$target$y_orig)/2),
  end = NROW(object$target$y_orig),
  h = 1,
  alpha = NULL,
  trace = FALSE,
  n_iter = 5000,
  ...
)
}
\arguments{
\item{object}{an object of class \dQuote{bsts.spec} or \dQuote{arima.spec}.}

\item{start}{numeric data index from which to start the backtest.}

\item{end}{numeric data index on which to end the backtest. The backtest will
end 1 period before that date in order to have at least 1 out of sample value
to compare against.}

\item{h}{forecast horizon. As the expanding window approaches the \dQuote{end},
the horizon will automatically shrink to the number of available out of sample
periods.}

\item{alpha}{optional numeric vector of coverage rates for which to calculate
the quantiles.}

\item{trace}{whether to show the progress bar. The user is expected to have
set up appropriate handlers for this using the \dQuote{progressr} package.}

\item{...}{not currently used.}

\item{n_iter}{number of MCMC iterations.}
}
\value{
A list with the following data.tables:
\itemize{
\item prediction : the backtest table with forecasts and actuals
\item metrics: a summary performance table showing metrics by
forecast horizon (MAPE, MSLRE, BIAS and MIS if alpha was not NULL).
}
}
\description{
Generates an expanding window walk forward backtest.
}
\note{
The function can use parallel functionality as long as the user has
set up a \code{\link[future]{plan}} using the future package.
}
