% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{tsconvert.bsts.estimate}
\alias{tsconvert.bsts.estimate}
\alias{tsconvert}
\title{Convert an estimated model to DLM}
\usage{
\method{tsconvert}{bsts.estimate}(
  object,
  to = "dlm",
  draw = "mean",
  burn = SuggestBurn(0.1, object$model),
  ...
)
}
\arguments{
\item{object}{tsconvert an object of class \dQuote{bsts.estimate}.}

\item{to}{class to convert to. Currently only DLM supported, but KFAS may be 
added in the future.}

\item{draw}{either an integer or \dQuote{mean}, after eliminating any burn 
draws, representing what part of the distribution to convert.}

\item{burn}{draws to burn.}

\item{...}{not currently used.}
}
\description{
Converts an estimated BSTS object to a DLM object.
}
