% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsts-model.R
\name{tsdecompose.bsts.estimate}
\alias{tsdecompose.bsts.estimate}
\alias{tsdecompose}
\alias{tsdecompose.bsts.predict}
\title{Model Decomposition}
\usage{
\method{tsdecompose}{bsts.estimate}(object, ...)

\method{tsdecompose}{bsts.predict}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{bsts.estimate} or \dQuote{bsts.predict}}

\item{...}{not currently used.}
}
\value{
A list of class \dQuote{tsmodel.distribution} representing the 
distribution of the state components of the model.
}
\description{
Decomposes the estimated model or prediction into its component
parts (states).
}
