\name{tsmetrics.bsts.estimate}
\alias{tsmetrics}
\alias{tsmetrics.bsts.estimate}
\alias{tsmetrics.arima.estimate}
\title{
Performance Metrics
}
\description{
Performance metrics from an estimated model.
}
\usage{
\method{tsmetrics}{bsts.estimate}(object, ...)
\method{tsmetrics}{arima.estimate}(object, ...)
}
\arguments{
\item{object}{
An object of class \dQuote{arima.estimate} or \dQuote{bsts.estimate}.
}
\item{\dots}{
Not currently used.
}
}
\author{
Alexios Galanos
}
